backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP
httpx==0.27.2
aiofiles==24.1.0

# AI SDKs（按需安装，注释掉不用的）
anthropic==0.39.0
openai==1.54.0
# zhipuai==2.1.5  # 智谱AI（按需）

# Browser automation（用于平台登录 Cookie 获取）
playwright==1.48.0

# Task queue
apscheduler==3.10.4

# Utils
loguru==0.7.2
tenacity==9.0.0
